python/requirements.txt
# robot control 
asyncio==3.4.3
evdev==1.3.0
future==0.18.2
iso8601==0.1.13
PyYAML==5.3.1
pyserial==3.5
# speaker 
google-speech==1.1.0
num2words==0.5.10
py-espeak-ng==0.1.8
pyttsx3==2.90